+++ EssentialFeed/EssentialFeed/Feed Cache/FeedStore.swift
@@ -1,0 +1,16 @@
+//
+//  FeedStore.swift
+//  EssentialFeed
+//
+//  Created by Stephen Davis on 2025-05-26.
+//
+
+import Foundation
+
+public protocol FeedStore {
+  typealias DeletionCompletion = (Error?) -> Void
+  typealias InsertionCompletion = (Error?) -> Void
+  
+  func deleteCachedFeed(completion: @escaping DeletionCompletion)
+  func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
