+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -36,11 +36,3 @@ public final class LocalFeedLoader {
     }
   }
 }
-
-public protocol FeedStore {
-  typealias DeletionCompletion = (Error?) -> Void
-  typealias InsertionCompletion = (Error?) -> Void
-  
-  func deleteCachedFeed(completion: @escaping DeletionCompletion)
-  func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
-}
